Add initial run() tests, use swappable FS
